Refactor: redistran
